fix: Build errors caused by unextracted component
